malloc

+++ 0x0B-malloc_free/0-create_array.c
@@ -17,7 +17,7 @@ char *create_array(unsigned int size, char c)
 
 	ar = malloc(size * sizeof(char));
 	i = 0;
-	while (i < size)
+	while (i < (int)size)
 	{
 		ar[i] = c;
 		i++;
